+++ sink-server/docker/sink.go
@@ -91,6 +91,10 @@ func (e *DockerEngine) newSink(deploymentID string, dbService string, pkg *pbsub
 		withEndpoint = "-e " + e.endpoint
 	}
 
+	withBuffer := ""
+	if serviceName == "clickhouse" {
+		withBuffer = "--undo-buffer-size=12"
+	}
 	startScript := []byte(fmt.Sprintf(`#!/bin/bash
 set -xeu
 
@@ -98,8 +102,8 @@ if [ ! -f /opt/subservices/data/setup-complete ]; then
     /app/substreams-sink-sql setup $DSN /opt/subservices/config/substreams.spkg %s && touch /opt/subservices/data/setup-complete
 fi
 
-/app/substreams-sink-sql run $DSN /opt/subservices/config/substreams.spkg --on-module-hash-mistmatch=warn %s
-`, withPostgraphile, withEndpoint))
+/app/substreams-sink-sql run $DSN /opt/subservices/config/substreams.spkg --on-module-hash-mistmatch=warn %s %s
+`, withPostgraphile, withBuffer, withEndpoint))
 	if err := os.WriteFile(filepath.Join(configFolder, "start.sh"), startScript, 0755); err != nil {
 		fmt.Println("")
 		return conf, motd, fmt.Errorf("writing file: %w", err)