+++ docs/release-notes/change-log.md
@@ -9,6 +9,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+### Unreleased
+
+* Fixed the 'local substreams alpha service serve' clickhouse deployment (was failing with a message regarding fork handling...)
+* Added some output-stream info to logs
+
 ## v1.3.1
 
 ### Server